pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.7",
    "fastapi>=0.125.0",
    "huggingface>=0.0.1",
    "huggingface-hub>=0.36.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.4",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "opencv-contrib-python>=4.12.0.88",
    "pdfplumber>=0.11.8",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pymupdf4llm>=0.2.7",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "streamlit>=1.52.2",
    "streamlit-pdf-viewer>=0.0.27",
    "uvicorn>=0.38.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "marker-pdf>=1.10.1",
    "sentence-transformers>=5.2.0",
    "transformers>=4.57.6",
    "accelerate>=1.12.0",
    "flagembedding>=1.3.5",
    "sqlalchemy>=2.0.46",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
